split off MainMenu from Header
